fix(maintner): renamed packages for golint errors
