+++ drghs-worker/maintnerd/api/internalapi/internal_api.go
@@ -1,4 +1,4 @@
-package internal_api
+package internalapi
 
 import (
 	"fmt"
